DormDDL.sql code
Example showing the DDL (data definition language or data description language) for a model database used to record information about dorm residents.  My decision was to only set the ‘ON DELETE CASCADE’ with the foreign key ‘dorm_id’ in the ‘room’ table and the ‘resides’ table, and refrain from implementing the foreign key with cascade delete on the resident table. This choice was for logical reasons. In a ‘on cascade
delete’, when a record in the parent table is deleted then the corresponding record in the child table are deleted as well. In this case, it is logically sound to delete the room when the dorm that the room is in gets deleted. However, if a room gets deleted, this does not mean that a resident is deleted, as the resident can still exist but may have been moved to a different room or dorm. It would be unfortunate to lose all the residents data, due to their room no longer existing, especially if they were moving to another room or dorm. The resides table has a composite primary key. This means that both, not only one, of these values are essential to the record in the table. Therefore, if one of them is deleted, then the relationship ‘resides’ would no longer be logically sound, thus the record should be deleted on cascade. Consequently, only the room and resides tables required “ON DELETE CASCADE”.